renaming container and moving to more common place
